convert ALlPokemon.test to TS; add props
